refactor(notebook-cloud): consume shared notebook state stores
